rename: spot -> place로 이름 변경 refactor: 타입 선언 별도 파일로 분리
