+++ frontend/src/app/_components/place/constants.ts
@@ -1,6 +1,6 @@
-import type { TPost } from './Spot';
+import type { TPlace } from './types';
 
-export const postData: TPost[] = [
+export const placeData: TPlace[] = [
   {
     id: '1',
     region: '강원',
