+++ frontend/src/app/_components/place/index.ts
@@ -1,0 +1,4 @@
+import Place from './Place';
+import PlaceList from './PlaceList';
+
+export { Place, PlaceList };
