+++ frontend/src/app/_components/place/types.ts
@@ -1,0 +1,21 @@
+type TRegion =
+  | '강원'
+  | '경기'
+  | '경남'
+  | '경북'
+  | '전남'
+  | '전북'
+  | '제주'
+  | '충남'
+  | '충북';
+
+export interface TPlace {
+  id: string;
+  region: TRegion;
+  liked: boolean;
+  imageUrl: string;
+  title: string;
+  likeCount: number;
+  commentCount: number;
+  rating: number;
+}
